refactor: change video position v2
